merge separate manifest.json files
There were 2 divergences in the `manifest.json` files :
- `"applications"` block in the firefox manifest which specified the
  minimum supported version of the extension in the browser
- different icon sizes in the default `manifest.json`

This divergence seemed trivial to be maintained separately. Merging and
running the extension from the merged `manifest.json` file did not throw
any major errors in the running of the extension on Firefox as well as
Chromium.

It does not seem sensible to maintain 2 separate `manifest.json` files
for the extension unless there are other chromium-based browsers who use
the `"applications"` block specifically for something else or some other
browser that accepts only one icon in the manifest.