eliminate grunt-copy ( alias : grunt build )
Now that the only minor divergence in the source code have been merged,
the `src` directory can be used as the single source of truth for both
source code as well as distribution code. There is nothing that needs to
be compiled or transpiled that would require a separate dedicated
distribution directory.