eliminate grunt-compress ( alias : grunt package )
`npm` has a built-in `pack` command that can compress the repository (
or specific `files` in the repository ) and generate a tarball.
Additionally, it also adds some default files like the README, LICENSE
and CHANGELOG for posterity.
The output tarball is also aptly named in the format
`<project-name>-<version-string>.tgz`

reference : https://stackoverflow.com/a/55220535

While the tarball can be generated by running the `npm pack` command,
for sake of backward-compatibility of existing `npm` scripts, it has
been assigned to the `package` script replacing the `grunt package`
alias.